iio.py: improved handling of flattened ndarrays
This does not modify the API, just the way some bad inputs are handled.
More precisely, what happens when the user calls iio.display or
iio.gallery with vectors of size (h*w,) instead of arrays of (h,w).
This is a common error in interactive notebooks, when one forgets
to call .reshape(h,w).  Up to now, the behavior in this case was to show
an error, because the image cannot be displayed.

This commit introduces a crazy heuristic: if a huge 1D vector is
detected as input, and global variables of names "w", "h" happen to
exist, and their values are such that the length of the vector is w*h,
then the vector is .reshaped(h,w)

This only works inside jupyter notebooks, because global variables are
globally visible in there.